LayoutTests/inspector/canvas/updateShader-webgl-expected.txt

CONSOLE MESSAGE: WebGL: ERROR: 0:1: 'INVALID' : syntax error
CONSOLE MESSAGE: WebGL: ERROR: 0:1: 'INVALID' : syntax error
WebGL tests for Canvas.updateShader command.


== Running test suite: Canvas.updateShader.WebGL
-- Running test case: Canvas.updateShader.WebGL.Vertex.Valid

    void main(void) {
        gl_Position = vec4(1, 2, 3, 4);
    }


-- Running test case: Canvas.updateShader.WebGL.Fragment.Valid

    precision mediump float;

    void main(void) {
        gl_FragColor = vec4(0.1, 0.2, 0.3, 0.4);
    }


-- Running test case: Canvas.updateShader.WebGL.Vertex.Invalid
PASS: Should produce error.
Error: Failed to update shader of given shaderType for given programId

-- Running test case: Canvas.updateShader.WebGL.Fragment.Invalid
PASS: Should produce error.
Error: Failed to update shader of given shaderType for given programId

-- Running test case: Canvas.updateShader.WebGL.ShaderType.Invalid
PASS: Should produce an error.
Error: Unknown shaderType: INVALID_SHADER_TYPE
